feat: Add url_launcher package and implement launching preview link
This commit adds the `url_launcher` package as a dependency and implements launching the preview link of a book when the "Free preview" button is pressed in the `BookActions` widget.. Additionally, it includes updates to the `android` and `windows` directories to register the `url_launcher` plugin. Finally, it adds two new utility functions, `customSnackBar` and `launchCustomUrl`, to display a custom snackbar and launch a custom URL respectively.